feature: return errors for many function.
This changes many function signature to now return an error in addition to whatever they
previously were returning. It is a breaking change, but seems like needed in order to make
the package more robust and so callers can more easily determine what has gone wrong if/when
something does.

Signed-off-by: deadprogram <ron@hybridgroup.com>